requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
PyGithub==1.59.1
XlsxWriter==3.1.2
waitress==2.1.2
Werkzeug==2.3.7
pytz==2024.1
